+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## Changelog
-- sync `1.21.8-0.25.5-sakura.4`
-- update NeoForge version to `21.8.31`
+- fix about `GuiRenderer` mixins
+- change malilib stub icon